PolyBot/requirements.txt
# Layer 1: Data and Market Access
py-clob-client>=0.5.0
python-dotenv>=1.0.0

# Layer 2: AI Brain
anthropic>=0.40.0
httpx>=0.27.0

# Layer 3: Math Engine
numpy>=1.26.0

# Layer 4: Execution
web3>=6.15.0

# Layer 5: Monitoring and Alerts
aiogram>=3.4.0

# Layer 6: Infrastructure
aiosqlite>=0.20.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
